requirements.txt
fastapi==0.115.12
sqlalchemy==2.0.40
python-dotenv==1.1.0
uvicorn==0.34.1
aiokafka==0.12.0
asyncpg==0.30.0
redis==5.2.1
fastapi-cache2==0.2.2
prometheus_fastapi_instrumentator==7.1.0
pytest==8.3.5
pytest-asyncio==0.26.0
httpx==0.28.1
asgi-lifespan==2.1.0
alembic>=1.13